requirements.txt
asgiref==3.8.1
cryptography-fernet==0.1.0
Django==5.1.3
html5lib==1.1
mechanize==0.4.10
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
webencodings==0.5.1
